fix migrate cli commands

+++ packages/lib/src/cli/generateTsConfig.ts
@@ -3,24 +3,19 @@ import fs from 'fs-extra'
 
 const base = require('./tsConfig.json')
 
-function generateTsConfig(buildPath: string): string {
+function generateTsConfig(
+  buildPath: string,
+  skipIfExists: boolean = false,
+): string {
   const modified = { ...base }
-
-  // const existingAbsolutePaths = modified.include.map(p => {
-  //   return path.join(__dirname, p)
-  // })
-
   const buildAbsolutePath = path.resolve(buildPath)
-
-  // modified.include = [...existingAbsolutePaths, buildAbsolutePath]
   modified.include = [buildAbsolutePath]
-
-  // modified.exclude = modified.exclude.map(p => {
-  //   return path.join(__dirname, p)
-  // })
 
   const filename = 'generatedTsConfig.json'
   const filePath = path.join(process.cwd(), filename)
+
+  if (skipIfExists && fs.existsSync(filePath)) return filePath
+
   fs.writeFileSync(filename, JSON.stringify(modified, null, 2))
   return filePath
 }
